scripts/requirements.txt
supabase==2.3.0
requests==2.31.0
psycopg2-binary==2.9.10
